backend/requirements.txt
# All_Chat - Python Dependencies
# Pin major versions for reproducibility; update deliberately.

# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12       # form/file upload parsing

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0                # async PostgreSQL driver
alembic==1.14.0                # database migrations

# Cache / rate limiting
redis==5.2.0

# Auth & security
python-jose[cryptography]==3.3.0   # JWT
argon2-cffi==23.1.0                # Argon2id password hashing
bleach==6.2.0                      # HTML sanitization
pydantic[email]==2.10.2            # request validation
pydantic-settings==2.6.1           # settings from env

# Email
aiosmtplib==3.0.2

# Images
Pillow==11.0.0

# HTTP client (link previews)
httpx==0.28.0
beautifulsoup4==4.12.3
lxml==5.3.0                        # fast BS4 parser

# Cryptography
cryptography==43.0.3               # AES-GCM, X25519

# Post-quantum (optional — install separately if liboqs is available)
# liboqs-python==0.11.0            # Kyber-768; requires liboqs C library

# Config
python-dotenv==1.0.1
